Fix many errors and failures found by unittests.
Many filter classes gain a _verify_children method which overrides
the default 'do-nothing' version.

All the remaining errors and failures, along with related passes, are
moved from test_filters to files named for the filter tested.

A small PGN file is added for the cases where the cql run has to
evaluate the query rather than just parse it.

An assumption is made on the location of the default output PGN file
which may be created, in an attempt to tidy up by deleting it.