Rewrite parsing of '--' and '[x]' filters.
The code in this commit is not a complete fix.

Existing unittests for '--' and '[x]' are passed now but some failed
previously.

Some CQL-6.2 examples which passed parsing before now fail.  In the
excelsior-return example making the precedence of the '--' and '[x]'
filters higher, not lower, than the 'find' filter allows the example to
pass parsing.  The dash filter documentation states 'A -- filter has
extremely low precedence' and it seems setting the precedence of '--'
lower than any other filter is too extreme.  Fixing this is left to
another commit.