Avoid repeating items generated for '--' filter in function body.
Queries like 'function F(T){--=T(R)}F(<T>)', varying the role of 'T'
within the body for argument '<T>', are tested.

One pattern, '=T' with '<T>' a quoted string, always gives ERROR as the
unittest outcome.  This seems not relevant to the purpose of the change
because 'function F(){--="Q"}F()' and '--="Q"' give the same ERROR.