Fix comparison of 'typename' and 'piecename' filters.
Three of the four CQL-6.2 examples affected do not use the 'path' filter
so the claim in an earlier commit that the known problems are related
to the 'path' filter was wrong.

All CQL-6.0.4, CQL-6.1, and CQL-6.2, examples now pass parsing.

About twenty filters may still be affected by this problem, the 'tag'
filter for example, but since no CQL examples fail parsing fixing any
other cases of this problem is outside the scope of this commit.

Version 3.0.dev5 of chessql.