Enable 'implicit search parameter' within 'str', and similar, filters.
The similar filters are 'message', 'comment', and '///'.  The enabling
affects '///' and the parenthesized versions of 'comment', 'message',
and 'str'.

The problem with a '///' at the end of the input stream is outside the
scope of this commit.

Replicating CQL validation of the FEN-like string in 'fen "<FEN>"' is
outside the scope of this commit.